noob-vs-pro-god-tasks/pro_research_crew/pyproject.toml
[project]
name = "pro_research_crew"
version = "0.1.0"
description = "pro_research_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "agentops>=0.3.21",
    "crewai[tools]>=0.86.0,<1.0.0",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
]
[project.scripts]
pro_research_crew = "pro_research_crew.main:run"
run_crew = "pro_research_crew.main:run"
train = "pro_research_crew.main:train"
replay = "pro_research_crew.main:replay"
test = "pro_research_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
